requirements.txt
opencv-python
mediapipe 
numpy
pyqt
